refactor: import pure helpers from testDialsPrompt.js in GenerateTestModal
